Match integer types throughout function body.
Rust DOES NOT automatically convert between one type and another
type.

In this exercise, `multiplier` was an unsigned 8-bit int being
used in an arithmetic expression of unsigned 32-bit ints.

So, one solution is to change the type of `multiplier` to match
`a` and `b`.

Note:  This exercise emphasizes that even *lossless* conversions
are not automatically performed. That is, even though `u8` could
fit in the memory allocated for `u32`, there is no automatic type
conversion.