Define variable using function args.
The compiler can infer a type for a variable based on the context
that the variable is used.

In this exercise, `distance` does not need its type annotated
because the compiler can infer its type as `u32` based on the
context `distance` is being used.

Specifically, `distance` is instantiated using the function args,
which the compiler already knows are `u32`.