Return value for each case of `n`
This exercise limits solutions to tools from previous chapters.

In this exercise, there are many possible values for `n` and each
value must map to a return value.

This scenario is the same as in  the `if-else` problem, solved
earlier.

So, `if-else` is the concept used to solve this problem.