Refactor factorial function
This exercise is primarily focused on mutability. Requiring the
use of `while` prompts the compiler to point out which variables
need mutability.

The `while` loop body shows that the condition variable, `n`, is
being *changed* during each iteration of the loop.

Since variables are immutable by default, `n` is changed to
`mut n` in the function signature.

This solution is an example of an accumulator pattern.
So, the accumulator variable `result` has similar requirements
on mutability because it also changes during each iteration.