Refactor `factorial` function without using counter
In the previous exercise, the factorial is calculated with a `while`
loop.

The decision to use `while` forced the program to include a counter
variable that keeps track of the number of iterations (recall
`n -= 1`).

The `for` loop automatically handles counting the number of
iterations, given some range.

In this exercise, the range is inclusive because the final range
value (5) must also contribute to the result.