Update dev profile to wrap value on int overflow
Profile settings are coupled to different cargo commands. For
example, the `dev` profile is used when calling `cargo test`.

The solution to the exercise disables overflow checking so the
program doesn't panic (crash) when an integer overflow happens, like
with `factorial(20)`.

So, for this exercise, wrapping is sensible. But for other exercises,
it's preferrable to crash than to produce (silently) wrong results -
it's difficult to fix a problem that you don't even know exists!

As understood, `overflow-check` should be re-enabled after testing.