Add conversion tests between several primitive types
The key to completing this exercise is thinking about the values
as they are represented in memory.

In the first test, a value is converted from a smaller data type
to a larger data type.

This scenario is the recommended use case for infallible
conversions using `as` because the memory allocated for the
larger `u32` is more than the memory required to store the
smaller `u16`.

In the next test, the amount of allocated memory does not change
as the type is converted. Instead, the meaning of the bits changes.
Memory is allocated to store the value `255` as theese eight bits:
`11111111`.

Signed integers use one bit (e.g. most significant bit) to carry
the sign information and the remaining bits to compute the value.
The MSB of `11111111` is 1, so the answer must be negative.
Since the answer is negative, the remaining bits are a two's-compliment
binary number. So allowing for overflow, `255 as i8` is `-1`.

These same concepts can be applied to the final test.