Add validation to `Ticket` instantiation
Validation is essentially checking *if* certain conditions are
true.
So, we can use `if-else` branching learned in a previous chapter
to scaffold a solution.

Validating `title` is similar to validating `description`.
This solution uses `if` and `else if` to ensure that each parameter
is checked independently.

Consequently, the error messages can be written to specify the
parameter causing the panic.