Make `Ticket` available in `helper` scope
The compiler can solve this exercise.
Before any changes are made to the exercise source code, run
`cargo test` and the compiler throws an error.
This error says `Ticket` is not in scope of `helper`
The compiler suggests adding `crate::Ticket` to resolve the error.

This suggestion makes sense because the `Ticket` struct is defined
in the crate root, `lib.rs`.
To build a path starting from the crate root this path structure
can be used: `use crate::<item to bring into scope>`.