Add setter methods
There are two common ways to write setter methods.
One way is to take `self` as a function parameter.
This method is good when you want to chain together setter methods,
one after another.
But since these methods consume the value `self`, you have to
remember to shadow the caller when using this method.

An alternative way is to take `&mut self` as a function parameter.
Doing this means that you can't chain the methods together, but
the code is imperative; that is, nothing gets returned from these
kinds of setters.
Instead, the field is mutated in place.

In the `works()` function of the test module, example usage of the
setters is given.
Since there is no variable shadowing happening here, setters are
added that take `&mut self`.