Add size in bytes for each type
This exercise can be solved by understanding the discussion on
`std::mem::size_of` for the `u8` datatype from the reading.
The example asserts that the answer is "1" and we know, by now,
that a `u8` occupies 8 bits in memory.
So, we can reasonably conclude that `size_of` "counts" the the
number of *bytes*.

The syntax to use the `size_of` function is different from what
we've encountered, so far.
`size_of` returns the size of a type, *in general*.
So, we say that `size_of` is a generic function.
To use it, we add the `::<foo_type>` syntax, before the parenthesis.
Generic functions are discussed in an upcoming section on
traits.
Generic functions let us write flexible source code, while also
satisfying the strong type system of Rust.
In the `tests` module, the assertions indicate exactly on which
type  we're interested in calling this function, for each test.