Add correct size for each test
One way to solve this problem is to let the compiler give you the answer.
We know that a `String` takes up a non-zero amount of space.
So if we assert the opposite, the compiler will complain and tell us
exactly what the mismatch is.
```
assertion: left == right
left: 24
right: 0
```
This error gives us a clue about what's going on: these 24 bytes
match the three values stored on the stack for a `String`, the pointer,
its capacity, and length.

In the committed solution, the problem is solved more generally.
There are a total of 3 parts of a `String` stored on the stack, as
discussed above.
Then, we recall from the reading that these parts are stored as
integers based on the machine architecture.
Rust gives access to whatever this integer size is through `usize`.
Finally, the solution requires 3 of whatever the size of `usize` is.

The same strategies can be used to pass the second test.