Restrict generic parameter to implement partial ordering
Recall that the `<=` operator is shorthand for a method provided by a trait,
just like `+` is shorthand for `add()` from the `Add` trait.
Searching the documentation, we see that `<=` is shorthand for `le()`
from the `PartialOrd` trait.
A trait must be in scope to use its method(s).
Since `PartialOrd` is in the `prelude`, we don't have to include a `use` statement.

The compiler suggests restricting the generic type parameter `T` by specifying
that `T` implements `PartialOrd`.

NB: What other traits would work?