Refactor accessor methods to return string slice
Observe that the unit tests still pass after the refactor.
Somehow, the compiler allows `&String` and `&str`  to be used interchangeably.

A `String` has memory allocated on the heap for the textual data itself and meta-
data stored on the stack.
This meta-data includes a pointer that points to the heap allocated textual data.
So, a `&String` is a different pointer whose head points to the pointer of the `String`
meta-data.
A string slice `&str` is also a pointer, but its head points to the beginning of a
substring of the heap allocated textual data.