Normalize getter output
The key to this exercise is recognizing that the necessary method is
found in the documentation for `str`, not `String`.
Searching the documentation for "String" vs. "str" shows there are
two separate modules in the Rust standard library.
The `std::string` module does not contain any functions that trim
the whitespace from the beginning and end of text.
On the other hand, `std::str` has the `trim()` method that does.