Add infallible conversions between `WrappingU32` and `u32`.
The important part of this exercise is practicing how to implement
traits.
Move on *iff* this exercise is solvable without using a reference to
look up the syntax for implementing the `From` trait.

The implementation of the `Into` trait, seen in this chapter, uses
generics in a way we have not yet seen.
```
impl <T, U> Into<U> for T ...
//   ^^^^^^
```
Removing the highlighted parameters makes the compiler complain
because these parameters are not *in the scope* of the `impl` block.
So, this signature tells the compiler that there are two generic parameters
in this scope.