Define custom exponentiation implementation for `u32`
Exponentiation is a binary *operation*, just like `Add`ing.
So, the structure of a solution may look like the implementation
for the `Add` trait.
This observation provides a scaffold for the syntax used in this commit.
Next, we rename the generic type parameter to make it clear that `n` is
an `Exponent`, just like the generic parameter in `Add` is named `RHS`.
Based on the unit tests, the conversions of interest are `u32` to many.
So, the associated type should be `u32` and this fact is reflected in the
function signature.
`u32` types have a `pow` method that calculates the output.
So, we can use that method inside the implementation block.

An alternative solution (and my initial attempt) calculated the exponent
with an accumulator pattern.
The present solution is shorter and more readable.