Add `Clone` trait to `Ticket`
The public `summary()` function owns `ticket`
that is passed in as a function parameter.
In the body of this public function, we call
a summary method on this instance: `ticket.summary()`.
So, a use-after-move situation is created.

The `Clone` trait unlocks deep copying for values
that implement this trait.
Deriving `Clone` implements cloning for `Ticket`.
So when the compiler reaches the call to the
instance method, it will make a deep copy of
`ticket` and call the `summary` method on this
deep copy.

In the final solution, the `summary()` method
is called on the function parameter binding
of `ticket` and we create a deep copy at the
other element of the tuple using the `clone`
method.