Implement the `Add` trait for `WrappingU32`
The test case we want to cover is adding together `WrappingU32`s.
Recall that `+` is syntactic sugar for the `add()` method from
`std::ops::Add`.
To use a trait method on a type, the type must *implement* the
trait and the trait must be in scope.
So, at least, we need to implement `Add` for the type after
bringing it into scope.

Implementing `Add` specifies an associated type.
According to the `assert` macro, we also want the return type
to be `WrappingU32`.
So, we assign the associated type the `WrappingU32` type and
also set it as the function's return type.

The test covers the case where more is added to a `u32` that
is already at its maximum value.
But, the assertion shows that the answer is significantly less
than the maximum for `u32`s.
So, we need a function that adds and wraps on integer
overflow: `wrapping_add`.

NB: I also found that `overflowing_add` works, if you know how
to destructure a tuple.

From this point, the compiler will suggest implementing some
missing traits.
Specifically, `Debug` and `PartialEq` are required because
the `assert!(...)` macro uses methods from these traits.
We *want* the `Copy` trait so the `add()` method
does not consume the underlying `u32`.
Since `Copy` is a subtrait of `Clone`, we *must* implement
the `Clone` trait, too.
Each of these traits is derivable, so just annotate `WrappingU32`.