Implement common traits for `SaturatingU16` type
This exercise can be broken down into sub-problems based on the
trait being implemented.
At almost every step, this solution takes advantage of symmetry
within the sub-problem to easily implement the trait for certain
cases (i.e. borrowed cases).

Since `SaturatingU16` already holds a `u16` value, converting
`From<u16>` is a matter of binding the data to the `value` field.
Converting `From<u8>` is the same problem, but we need to turn
`u8` into `u16`, first, with `into()`.
Solving the borrowed (`&`) versions of these sub-problems adds
the extra step of dereferencing (`*`), before relying on the
original `From` implementations.

Implementing the `Add` trait is done similarly.
In this case, `Add<SaturatingU16> for SaturatingU16` and `Add<u16>
for SaturatingU16` are the base cases.
To satisfy the condition that addition saturates at the value for
`u16`, these base cases use the `saturating_add()` function.

The borrowed versions dereference the values first, then use `+` to
add values together.
It's possible to just use `+` here because `+` is syntactic sugar
for `Add::add()`, which was implemented for the dereferenced types,
above.

Finally, this exercise requires that we use different methods to
implement the `PartialEq` trait.
Deriving `PartialEq` unlocks comparison with other `SaturatingU16`.
Writing a custom implementation of `PartialEq::eq` also unlocks
comparison with different types.
This solution implements `PartialEq<u16>` based on the example in the standard library documentation.