Refactor assignee accessor method to optionally return `Some` or `None`
Before this commit, `assigned_to` would panic at runtime if the `Status`
was `ToDo` or `Done`.

This commit refactors `assigned_to` to optionally return a value.
Now, the source code models that there's nothing to return if the `Status`
was `ToDo` or `Done`.

Also, this commit shows two ways to refactor the method.
The first approach takes advantage of `match` exhaustiveness by explicitly
handling each `Status` variant.
Prefer this approach because, if a new variant is added to `Status`, the
compiler will point out where in the source code the new variant
must be handled.

The second approach uses `if let`.
This approach technically works, but if a new variant is added, the
compiler will not point out the areas that need to be modified.