Add ticket creation convenience method
`easy_ticket` returns a `Ticket`.
There's only one way to create a ticket; `Ticket::new()`.
So, the solution (body of `easy_ticket`) must include a call to `Ticket::new()`.
This static method returns a `Result`, so the two options are `match` or `unwrap`.
`unwrap`-ing here is wrong because `new()` can return an `Err` in multiple ways.
Returning `Err` causes `unwrap` to panic.
But, we only want to panic when the title is invalid, not in every `Err` case.
So, we match on the `Result<Ticket, String>`.
Inside the `Err` branch, we check if the error message `String` includes
"Description" and respond accordingly.