Refactor convenience method to explicitly handle `TicketNewError`
This commit adds an error `enum` to model invalid titles and descriptions.
When the description is invalid, we can ignore the error message with `_`.

In this problem, emphasis is on how much less code we have to write to add
the same error handling functionality as the previous exercise,
`05_ticket_v2/07_unwrap/`.