Create new library
To bring attention to the `Cargo.toml`, the package name is changed to
`problem_library`.
The `package` `name` is the start of a path to content inside `lib.rs`.
So following from this optional step, we update the `use` statement in `main.rs`.
The compiler tells us that there's an unresolved import for `hello_world()`.
We know from the `use` statement, that the compiler is expecting `hello_world()`
in `lib.rs`, so we add it as a public function.