Add fallible conversions for `Status`
Before this commit, `Status` modeled the variety of valid `Status` states.
This model could be matched on, such as when `Ticket` creation was refactored.
But otherwise `Status` did not have behavior attached to it.
As a result, the library did not have a way to make an instance of
a particular `Status` variant.

In this commit, the `TryFrom` trait is implemented for `Status`.
Now, it's possible to make an instance of a particular variant by converting
from a `String` (or `&str`) representation to a `Status`.